Add, Edit added

+++ backend/src/main/java/summer23/backend/web/ArtistController.java
@@ -1,5 +1,7 @@
 package summer23.backend.web;
 
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -7,6 +9,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import org.springframework.web.servlet.view.RedirectView;
 
 import summer23.backend.domain.Artist;
 import summer23.backend.domain.ArtistRepository;
@@ -28,8 +32,15 @@ public class ArtistController {
 	@RequestMapping(value = "/addartist")
 	public String addArtist(Model model){
 		model.addAttribute("artist", new Artist());
-		return "addArtist";
+		return "addartist";
 	}
+
+	// // Search for existing artists
+    // @GetMapping("/searchartist")
+    // public String searchArtist(@PathVariable("name") String artistName, Model model) {
+    //     model.addAttribute("artists", artistRepository.findByName(artistName));
+    //     return "addsong";
+    // }
 	
 	// Save new artist
 	@PostMapping("/saveartist")
@@ -37,13 +48,20 @@ public class ArtistController {
 		artistRepository.save(artist);
 		return "redirect:artistlist";
 	}
-	
-	// Delete artist
+
+	// Delete artist, if there's no songs
 	@GetMapping("/deleteartist/{id}")
-	public String deleteArtist(@PathVariable("id") Long artistId, Model model){ 
-		artistRepository.deleteById(artistId);
-		return "redirect:../artistlist";
-	}
+    public RedirectView deleteArtist(@PathVariable("id") Long artistId, RedirectAttributes attributes) {
+        Optional<Artist> artist = artistRepository.findById(artistId);
+        // From optional to normal
+        Artist normalArtist = artist.get();
+        if (normalArtist.getSongs().isEmpty()) {
+            artistRepository.deleteById(artistId);
+        } else {
+            attributes.addFlashAttribute("error", "Artist can't be deleted since it has songs!");
+        }
+        return new RedirectView("/artistlist");
+    }
 	
 	// Edit artist
 	@GetMapping("/editartist/{id}")
