+++ backend/src/main/java/summer23/backend/web/SongController.java
@@ -1,5 +1,7 @@
 package summer23.backend.web;
 
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -7,7 +9,10 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import org.springframework.web.servlet.view.RedirectView;
 
+import summer23.backend.domain.Artist;
 import summer23.backend.domain.ArtistRepository;
 import summer23.backend.domain.GenreRepository;
 import summer23.backend.domain.Song;
@@ -55,12 +60,19 @@ public class SongController {
 		return "redirect:songlist";
 	}
 	
-	// Delete song
+	// Delete Song, if there's sheet music
 	@GetMapping("/deletesong/{id}")
-	public String deleteSong(@PathVariable("id") Long songId, Model model){ 
-		songRepository.deleteById(songId);
-		return "redirect:../songlist";
-	}
+    public RedirectView deleteSong(@PathVariable("id") Long songId, RedirectAttributes attributes) {
+        Optional<Song> song = songRepository.findById(songId);
+        // From optional to normal
+        Song normalSong = song.get();
+        if (normalSong.getNotes().isEmpty()) {
+            songRepository.deleteById(songId);
+        } else {
+            attributes.addFlashAttribute("error", "Song can't be deleted since it has sheet music!");
+        }
+        return new RedirectView("/songlist");
+    }
 	
 	// Edit song
 	@GetMapping("/editsong/{id}")
